Move the `PredictorStream` from `src/core/stream.js` and into its own file
